Simple PDF embed code
With this its possible to embed a PDF document on a webpage if the
documents-folder is located in the same folder as this file and the
PDF-file (test.pdf) inside documents.  No certainty if this can be
configured to work with a server and database.

Works at least with Chrome and has built-in zoom in and out, as well as
other options.