Folder containing barebones form for document upload
Included are the controller (UploadController.php) and view template
(Upload/index.ctp) that goes in the Template folder. Error will happen
when pressing the upload-button, I havent figured out how to fix it even
with the model where the function is. The page is accessed by typing
localhost/TestBedForDigitalMetadata/upload in the browser addressbar
when logged in as admin.